feat: 채팅 참여자 목록에서 프로필 클릭
- 이미지 업로드 & 프로필 보기 시,  변경이 필요
  -> getAvatar(): Profile.tsx 에서 ProfileData.tsx로 이동
- 이미지 업로드 시 오른쪽 내 프로필 컴포넌트 업데이트 하려면?
  -> context 이용 필요?? 더 고민 해보

#71
